% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordis.R
\name{cordis_con}
\alias{cordis_con}
\title{Connection to CORDIS database}
\usage{
cordis_con()
}
\value{
a database connection
}
\description{
This function returns a connection to the CORDIS database
}
\details{
the database location is at the cachedir/cordisdb
}
\examples{
\dontrun{
library(dplyr)
con <- cordis_con()
con \%>\% tbl("projectPublications")
}
}
