% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordis.R
\name{cordis_disconnect}
\alias{cordis_disconnect}
\title{Close and shutdown CORDIS database connection}
\usage{
cordis_disconnect(con)
}
\arguments{
\item{con}{database connection to duckdb database}
}
\value{
invisible TRUE on success
}
\description{
This function closes a connection to the CORDIS database
}
\details{
also takes care of shutting down the connection
}
\examples{
\dontrun{
con <- cordis_con()
cordis_disconnect(con)
}
}
