% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordis.R
\name{cordis_schema}
\alias{cordis_schema}
\title{CORDIS schema}
\usage{
cordis_schema()
}
\value{
a dataframe with database schema metadata
}
\description{
Description of all tables, their fields and data types
}
\examples{
\dontrun{
cordis_schema()
}
}
