% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordis.R
\name{cordis_tables}
\alias{cordis_tables}
\title{CORDIS tables}
\usage{
cordis_tables()
}
\value{
a dataframe with table names and row counts
}
\description{
Enumeration of tables present in CORDIS database
}
\examples{
\dontrun{
cordis_tables()
}
}
